data-processing/requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
flopy==3.9.3
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.2
six==1.17.0
tzdata==2025.2
psycopg2-binary==2.9.9
